+++ codec/cbor/codec.go
@@ -4,6 +4,7 @@ import (
 	"reflect"
 
 	"github.com/dogmatiq/marshalkit/codec"
+	"github.com/dogmatiq/marshalkit/codec/internal/portablename"
 	"github.com/fxamacker/cbor/v2"
 )
 
@@ -20,7 +21,7 @@ func (Codec) Query(types []reflect.Type) codec.Capabilities {
 	}
 
 	for _, rt := range types {
-		if n, ok := portableName(rt); ok {
+		if n, ok := portablename.FromReflect(rt); ok {
 			caps.Types[rt] = n
 		}
 	}
@@ -42,19 +43,4 @@ func (Codec) Marshal(v any) ([]byte, error) {
 // Unmarshal decodes a binary representation into v.
 func (Codec) Unmarshal(data []byte, v any) error {
 	return cbor.Unmarshal(data, v)
-}
-
-// portableName returns the portable name to use for the given type.
-func portableName(rt reflect.Type) (string, bool) {
-	n := rt.Name()
-	if n != "" {
-		return n, true
-	}
-
-	for rt.Kind() == reflect.Ptr {
-		rt = rt.Elem()
-	}
-
-	n = rt.Name()
-	return n, n != ""
 }