+++ codec/internal/portablename/doc.go
@@ -1,0 +1,2 @@
+// Package portablename generates portable names for arbitrary types.
+package portablename