+++ codec/marshaler_test.go
@@ -3,7 +3,7 @@ package codec_test
 import (
 	"reflect"
 
-	. "github.com/dogmatiq/dogma/fixtures"
+	. "github.com/dogmatiq/enginekit/enginetest/stubs"
 	"github.com/dogmatiq/marshalkit"
 	. "github.com/dogmatiq/marshalkit/codec"
 	. "github.com/dogmatiq/marshalkit/codec/internal/fixtures"
@@ -24,9 +24,9 @@ var _ = Describe("type Marshaler", func() {
 		var err error
 		marshaler, err = NewMarshaler(
 			[]reflect.Type{
-				reflect.TypeOf(&ProtoMessage{}),
-				reflect.TypeOf(MessageA{}),
-				reflect.TypeOf(MessageB{}),
+				reflect.TypeFor[*ProtoMessage](),
+				reflect.TypeFor[CommandStub[TypeA]](),
+				reflect.TypeFor[CommandStub[TypeB]](),
 			},
 			[]Codec{
 				&protobuf.DefaultNativeCodec,
@@ -41,7 +41,7 @@ var _ = Describe("type Marshaler", func() {
 		It("returns an error if multiple codecs used the same media-type", func() {
 			_, err := NewMarshaler(
 				[]reflect.Type{
-					reflect.TypeOf(MessageA{}),
+					reflect.TypeFor[CommandStub[TypeA]](),
 				},
 				[]Codec{
 					&json.Codec{},
@@ -56,8 +56,8 @@ var _ = Describe("type Marshaler", func() {
 		It("excludes types with conflicting portable type names", func() {
 			marshaler, err := NewMarshaler(
 				[]reflect.Type{
-					reflect.TypeOf(&ProtoMessage{}),
-					reflect.TypeOf(&conflicting.ProtoMessage{}),
+					reflect.TypeFor[*ProtoMessage](),
+					reflect.TypeFor[*conflicting.ProtoMessage](),
 				},
 				[]Codec{
 					&json.Codec{},
@@ -81,8 +81,8 @@ var _ = Describe("type Marshaler", func() {
 		It("returns an error if types with conflicting portable type names are excluded by all codecs", func() {
 			_, err := NewMarshaler(
 				[]reflect.Type{
-					reflect.TypeOf(MessageA{}),
-					reflect.TypeOf(&MessageA{}),
+					reflect.TypeFor[CommandStub[TypeA]](),
+					reflect.TypeFor[*CommandStub[TypeA]](),
 				},
 				[]Codec{
 					&json.Codec{},
@@ -90,10 +90,10 @@ var _ = Describe("type Marshaler", func() {
 			)
 			Expect(err).To(Or(
 				MatchError(
-					"naming conflicts occurred within all of the codecs that support the 'fixtures.MessageA' type",
+					"naming conflicts occurred within all of the codecs that support the 'stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]' type",
 				),
 				MatchError(
-					"naming conflicts occurred within all of the codecs that support the '*fixtures.MessageA' type",
+					"naming conflicts occurred within all of the codecs that support the '*stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]' type",
 				),
 			))
 		})
@@ -101,15 +101,15 @@ var _ = Describe("type Marshaler", func() {
 		It("returns an error if there are unsupported types", func() {
 			_, err := NewMarshaler(
 				[]reflect.Type{
-					reflect.TypeOf(&ProtoMessage{}),
-					reflect.TypeOf(MessageA{}),
+					reflect.TypeFor[*ProtoMessage](),
+					reflect.TypeFor[EventStub[TypeA]](),
 				},
 				[]Codec{
 					&protobuf.DefaultJSONCodec,
 				},
 			)
 			Expect(err).To(MatchError(
-				"no codecs support the 'fixtures.MessageA' type",
+				"no codecs support the 'stubs.EventStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]' type",
 			))
 		})
 	})
@@ -117,33 +117,33 @@ var _ = Describe("type Marshaler", func() {
 	Describe("func MarshalType()", func() {
 		It("returns the portable type name", func() {
 			n, err := marshaler.MarshalType(
-				reflect.TypeOf(MessageA{}),
+				reflect.TypeFor[CommandStub[TypeA]](),
 			)
 			Expect(err).ShouldNot(HaveOccurred())
-			Expect(n).To(Equal("MessageA"))
+			Expect(n).To(Equal("CommandStub[TypeA]"))
 		})
 
 		It("returns an error if the type is not supported", func() {
 			_, err := marshaler.MarshalType(
-				reflect.TypeOf(MessageC{}),
+				reflect.TypeFor[CommandStub[TypeC]](),
 			)
 			Expect(err).To(MatchError(
-				"no codecs support the 'fixtures.MessageC' type",
+				"no codecs support the 'stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeC]' type",
 			))
 		})
 	})
 
 	Describe("func UnmarshalType()", func() {
 		It("returns the reflection type", func() {
-			t, err := marshaler.UnmarshalType("MessageA")
+			t, err := marshaler.UnmarshalType("CommandStub[TypeA]")
 			Expect(err).ShouldNot(HaveOccurred())
-			Expect(t).To(Equal(reflect.TypeOf(MessageA{})))
+			Expect(t).To(Equal(reflect.TypeFor[CommandStub[TypeA]]()))
 		})
 
 		It("returns an error if the type name is not recognized", func() {
-			_, err := marshaler.UnmarshalType("MessageC")
+			_, err := marshaler.UnmarshalType("CommandStub[TypeC]")
 			Expect(err).To(MatchError(
-				"the portable type name 'MessageC' is not recognized",
+				"the portable type name 'CommandStub[TypeC]' is not recognized",
 			))
 		})
 	})
@@ -151,13 +151,13 @@ var _ = Describe("type Marshaler", func() {
 	Describe("func Marshal()", func() {
 		It("marshals using the first suitable codec", func() {
 			p, err := marshaler.Marshal(
-				MessageA{
-					Value: "<value>",
+				CommandStub[TypeA]{
+					Content: "<value>",
 				},
 			)
 			Expect(err).ShouldNot(HaveOccurred())
-			Expect(p.MediaType).To(Equal("application/json; type=MessageA"))
-			Expect(p.Data).To(Equal([]byte(`{"Value":"\u003cvalue\u003e"}`)))
+			Expect(p.MediaType).To(Equal(`application/json; type="CommandStub[TypeA]"`))
+			Expect(p.Data).To(Equal([]byte(`{"content":"\u003cvalue\u003e"}`)))
 
 			p, err = marshaler.Marshal(
 				&ProtoMessage{
@@ -179,9 +179,9 @@ var _ = Describe("type Marshaler", func() {
 		})
 
 		It("returns an error if the type is not supported", func() {
-			_, err := marshaler.Marshal(MessageC{})
+			_, err := marshaler.Marshal(CommandStub[TypeC]{})
 			Expect(err).To(MatchError(
-				"no codecs support the 'fixtures.MessageC' type",
+				"no codecs support the 'stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeC]' type",
 			))
 		})
 	})
@@ -189,17 +189,17 @@ var _ = Describe("type Marshaler", func() {
 	Describe("func MarshalAs()", func() {
 		It("marshals using the codec associated with the given media type", func() {
 			p, ok, err := marshaler.MarshalAs(
-				MessageA{
-					Value: "<value>",
+				CommandStub[TypeA]{
+					Content: "<value>",
 				},
 				[]string{
-					"application/json; type=MessageA",
+					`application/json; type="CommandStub[TypeA]"`,
 				},
 			)
 			Expect(err).ShouldNot(HaveOccurred())
 			Expect(ok).To(BeTrue())
-			Expect(p.MediaType).To(Equal("application/json; type=MessageA"))
-			Expect(p.Data).To(Equal([]byte(`{"Value":"\u003cvalue\u003e"}`)))
+			Expect(p.MediaType).To(Equal(`application/json; type="CommandStub[TypeA]"`))
+			Expect(p.Data).To(Equal([]byte(`{"content":"\u003cvalue\u003e"}`)))
 
 			p, ok, err = marshaler.MarshalAs(
 				&ProtoMessage{
@@ -246,20 +246,20 @@ var _ = Describe("type Marshaler", func() {
 
 		It("ignores unsupported media-types", func() {
 			p, ok, err := marshaler.MarshalAs(
-				MessageA{},
+				CommandStub[TypeA]{},
 				[]string{
-					"application/binary; type=MessageA",
-					"application/json; type=MessageA",
+					`application/binary; type="CommandStub[TypeA]"`,
+					`application/json; type="CommandStub[TypeA]"`,
 				},
 			)
 			Expect(err).ShouldNot(HaveOccurred())
 			Expect(ok).To(BeTrue())
-			Expect(p.MediaType).To(Equal("application/json; type=MessageA"))
+			Expect(p.MediaType).To(Equal(`application/json; type="CommandStub[TypeA]"`))
 		})
 
 		It("returns an error if the media-type is malformed", func() {
 			_, _, err := marshaler.MarshalAs(
-				MessageA{},
+				CommandStub[TypeA]{},
 				[]string{
 					"<malformed>",
 				},
@@ -281,7 +281,7 @@ var _ = Describe("type Marshaler", func() {
 
 		It("returns false if the media-type is not supported", func() {
 			_, ok, err := marshaler.MarshalAs(
-				MessageC{},
+				CommandStub[TypeC]{},
 				[]string{
 					"application/json; type=MessageC",
 				},
@@ -292,7 +292,7 @@ var _ = Describe("type Marshaler", func() {
 
 		It("returns false if the portable name in the media-type does not match the value's type", func() {
 			_, ok, err := marshaler.MarshalAs(
-				MessageA{},
+				CommandStub[TypeA]{},
 				[]string{
 					"application/json; type=MessageC",
 				},
@@ -304,7 +304,7 @@ var _ = Describe("type Marshaler", func() {
 		It("panics if no media-types are provided", func() {
 			Expect(func() {
 				marshaler.MarshalAs(
-					MessageA{},
+					CommandStub[TypeA]{},
 					nil,
 				)
 			}).To(PanicWith("at least one media-type must be provided"))
@@ -315,12 +315,12 @@ var _ = Describe("type Marshaler", func() {
 		It("unmarshals using the first suitable codec", func() {
 			v, err := marshaler.Unmarshal(
 				marshalkit.Packet{
-					MediaType: "application/json; type=MessageA",
+					MediaType: `application/json; type="CommandStub[TypeA]"`,
 					Data:      []byte("{}"),
 				},
 			)
 			Expect(err).ShouldNot(HaveOccurred())
-			Expect(v).To(Equal(MessageA{}))
+			Expect(v).To(Equal(CommandStub[TypeA]{}))
 
 			v, err = marshaler.Unmarshal(
 				marshalkit.Packet{
@@ -379,7 +379,7 @@ var _ = Describe("type Marshaler", func() {
 
 	Describe("func MediaTypesFor()", func() {
 		It("returns media types in order of codec priority", func() {
-			mt := marshaler.MediaTypesFor(reflect.TypeOf(&ProtoMessage{}))
+			mt := marshaler.MediaTypesFor(reflect.TypeFor[*ProtoMessage]())
 			Expect(mt).To(Equal([]string{
 				"application/vnd.google.protobuf; type=dogmatiq.marshalkit.fixtures.ProtoMessage",
 				"application/vnd.google.protobuf+json; type=dogmatiq.marshalkit.fixtures.ProtoMessage",
@@ -388,7 +388,7 @@ var _ = Describe("type Marshaler", func() {
 		})
 
 		It("returns an empty slice when given an unsupported message type", func() {
-			mt := marshaler.MediaTypesFor(reflect.TypeOf(&MessageC{}))
+			mt := marshaler.MediaTypesFor(reflect.TypeFor[*CommandStub[TypeC]]())
 			Expect(mt).To(BeEmpty())
 		})
 	})