+++ fixtures/doc.go
@@ -1,3 +1,5 @@
 // Package fixtures contains test fixtures for the types from the
 // github.com/dogmatiq/dogma/fixtures package.
+//
+// Deprecated: This package will be removed in a future release.
 package fixtures