+++ CHANGELOG.md
@@ -10,6 +10,13 @@ The format is based on [Keep a Changelog], and this project adheres to
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
+## [Unreleased]
+
+### Deprecated
+
+- The `fixtures` package is now deprecated and will be removed in a future
+  release.
+
 ## [0.7.6] - 2024-09-26
 
 ### Changed